# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "AI 模型提供商管理服务 API 文档\n提供模型提供商、模型信息和参数规则的查询接口",
        "title": "Genkit AI Service API",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/providers": {
            "get": {
                "description": "获取系统中所有可用的模型提供商列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "providers"
                ],
                "summary": "获取所有提供商列表",
                "responses": {
                    "200": {
                        "description": "成功返回提供商列表",
                        "schema": {
                            "$ref": "#/definitions/ProviderListDataResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/providers/{providerId}": {
            "get": {
                "description": "根据提供商ID获取详细信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "providers"
                ],
                "summary": "获取提供商详情",
                "parameters": [
                    {
                        "type": "string",
                        "example": "gemini",
                        "description": "提供商ID",
                        "name": "providerId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功返回提供商详情",
                        "schema": {
                            "$ref": "#/definitions/ProviderDataResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "提供商不存在",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/providers/{providerId}/models": {
            "get": {
                "description": "获取指定提供商的所有可用模型列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "providers"
                ],
                "summary": "获取提供商的模型列表",
                "parameters": [
                    {
                        "type": "string",
                        "example": "gemini",
                        "description": "提供商ID",
                        "name": "providerId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功返回模型列表",
                        "schema": {
                            "$ref": "#/definitions/ModelListDataResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "提供商不存在",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/providers/{providerId}/models/{modelId}": {
            "get": {
                "description": "获取指定提供商的指定模型的详细信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "providers"
                ],
                "summary": "获取模型详情",
                "parameters": [
                    {
                        "type": "string",
                        "example": "gemini",
                        "description": "提供商ID",
                        "name": "providerId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "gemini-1.5-flash",
                        "description": "模型ID",
                        "name": "modelId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功返回模型详情",
                        "schema": {
                            "$ref": "#/definitions/ModelDataResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "提供商或模型不存在",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
        "/providers/{providerId}/models/{modelId}/parameter-rules": {
            "get": {
                "description": "获取指定模型的所有参数配置规则",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "providers"
                ],
                "summary": "获取模型参数规则",
                "parameters": [
                    {
                        "type": "string",
                        "example": "gemini",
                        "description": "提供商ID",
                        "name": "providerId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "gemini-1.5-flash",
                        "description": "模型ID",
                        "name": "modelId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功返回参数规则列表",
                        "schema": {
                            "$ref": "#/definitions/ParameterRuleListDataResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "提供商或模型不存在",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "CredentialSchema": {
            "type": "object",
            "properties": {
                "credential_form_schemas": {
                    "description": "凭证表单配置列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/CredentialFormSchema"
                    }
                }
            }
        },
        "ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应代码",
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "description": "响应信息",
                    "type": "string",
                    "example": "请求参数错误"
                },
                "traceId": {
                    "description": "追踪ID（用于全链路追踪和问题排查）",
                    "type": "string",
                    "example": "trace-1729756800-a1b2c3d4"
                }
            }
        },
...
        "Model": {
            "type": "object",
            "properties": {
                "deprecated": {
                    "description": "是否已弃用",
                    "type": "boolean",
                    "example": false
                },
                "features": {
                    "description": "特性列表",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "tool-call",
                        "multi-tool-call",
                        "stream-tool-call"
                    ]
                },
                "label": {
                    "description": "多语言标签",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    },
                    "example": {
                        "en_US": "Gemini 1.5 Flash",
                        "zh_Hans": "Gemini 1.5 Flash"
                    }
                },
                "model": {
                    "description": "模型标识",
                    "type": "string",
                    "example": "gemini-1.5-flash"
                },
                "model_properties": {
                    "description": "模型属性",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ModelProperties"
                        }
                    ]
                },
                "model_type": {
                    "description": "模型类型（llm、tts、text_embedding等）",
                    "type": "string",
                    "example": "llm"
                },
                "parameter_rules": {
                    "description": "参数规则",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ParameterRule"
                    }
                },
                "pricing": {
                    "description": "定价信息",
                    "allOf": [
                        {
                            "$ref": "#/definitions/Pricing"
                        }
                    ]
                }
            }
        },
        "ModelDataResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应代码",
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "description": "响应数据",
                    "allOf": [
                        {
                            "$ref": "#/definitions/Model"
                        }
                    ]
                },
                "message": {
                    "description": "响应信息",
                    "type": "string",
                    "example": "success"
                },
                "traceId": {
                    "description": "追踪ID（用于全链路追踪和问题排查）",
                    "type": "string",
                    "example": "trace-1729756800-a1b2c3d4"
                }
            }
        },
        "ModelListDataResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应代码",
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "description": "响应数据",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Model"
                    }
                },
                "message": {
                    "description": "响应信息",
                    "type": "string",
                    "example": "success"
                },
                "traceId": {
                    "description": "追踪ID（用于全链路追踪和问题排查）",
                    "type": "string",
                    "example": "trace-1729756800-a1b2c3d4"
                }
            }
        },
        "ModelProperties": {
            "type": "object",
            "properties": {
                "context_size": {
                    "description": "上下文大小",
                    "type": "integer"
                },
                "mode": {
                    "description": "模式（chat、completion等）",
                    "type": "string"
                }
            }
        },
        "ModelTypeInfo": {
            "type": "object",
            "properties": {
                "position": {
                    "description": "位置文件路径",
                    "type": "string"
                },
                "predefined": {
                    "description": "预定义模型列表",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "ParameterRule": {
            "type": "object",
            "properties": {
                "default": {
                    "description": "默认值"
                },
                "help": {
                    "description": "帮助信息（多语言）",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "label": {
                    "description": "标签（多语言）",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "max": {
                    "description": "最大值"
                },
                "min": {
                    "description": "最小值"
                },
                "name": {
                    "description": "参数名称",
                    "type": "string",
                    "example": "temperature"
                },
                "options": {
                    "description": "选项列表",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "required": {
                    "description": "是否必填",
                    "type": "boolean",
                    "example": false
                },
                "type": {
                    "description": "类型",
                    "type": "string",
                    "example": "float"
                },
                "use_template": {
                    "description": "使用的模板",
                    "type": "string",
                    "example": "temperature"
                }
            }
        },
        "ParameterRuleListDataResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应代码",
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "description": "响应数据",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ParameterRule"
                    }
                },
                "message": {
                    "description": "响应信息",
                    "type": "string",
                    "example": "success"
                },
                "traceId": {
                    "description": "追踪ID（用于全链路追踪和问题排查）",
                    "type": "string",
                    "example": "trace-1729756800-a1b2c3d4"
                }
            }
        },
        "Pricing": {
            "type": "object",
            "properties": {
                "currency": {
                    "description": "货币",
                    "type": "string"
                },
                "input": {
                    "description": "输入价格",
                    "type": "string"
                },
                "output": {
                    "description": "输出价格",
                    "type": "string"
                },
                "unit": {
                    "description": "单位",
                    "type": "string"
                }
            }
        },
        "Provider": {
            "type": "object",
            "properties": {
                "background": {
                    "description": "背景色",
                    "type": "string",
                    "example": "#4285F4"
                },
                "configurate_methods": {
                    "description": "配置方法列表",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "predefined-model",
                        "customizable-model"
                    ]
                },
                "help": {
                    "description": "帮助信息",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ProviderHelp"
                        }
                    ]
                },
                "icon_large": {
                    "description": "大图标（多语言）",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "icon_small": {
                    "description": "小图标（多语言）",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "id": {
                    "description": "提供商ID（从文件夹名称获取）",
                    "type": "string",
                    "example": "gemini"
                },
                "label": {
                    "description": "多语言标签",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    },
                    "example": {
                        "en_US": "Google Gemini",
                        "zh_Hans": "谷歌 Gemini"
                    }
                },
                "model_credential_schema": {
                    "description": "模型凭证配置",
                    "allOf": [
                        {
                            "$ref": "#/definitions/CredentialSchema"
                        }
                    ]
                },
                "models": {
                    "description": "模型类型配置",
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/ModelTypeInfo"
                    }
                },
                "provider": {
                    "description": "提供商标识",
                    "type": "string",
                    "example": "gemini"
                },
                "provider_credential_schema": {
                    "description": "提供商凭证配置",
                    "allOf": [
                        {
                            "$ref": "#/definitions/CredentialSchema"
                        }
                    ]
                },
                "supported_model_types": {
                    "description": "支持的模型类型列表",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "llm",
                        "text-embedding"
                    ]
                }
            }
        },
        "ProviderDataResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应代码",
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "description": "响应数据",
                    "allOf": [
                        {
                            "$ref": "#/definitions/Provider"
                        }
                    ]
                },
                "message": {
                    "description": "响应信息",
                    "type": "string",
                    "example": "success"
                },
                "traceId": {
                    "description": "追踪ID（用于全链路追踪和问题排查）",
                    "type": "string",
                    "example": "trace-1729756800-a1b2c3d4"
                }
            }
        },
        "ProviderHelp": {
            "type": "object",
            "properties": {
                "title": {
                    "description": "帮助标题（多语言）",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "url": {
                    "description": "帮助链接（多语言）",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
        "ProviderListDataResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "响应代码",
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "description": "响应数据",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Provider"
                    }
                },
                "message": {
                    "description": "响应信息",
                    "type": "string",
                    "example": "success"
                },
                "traceId": {
                    "description": "追踪ID（用于全链路追踪和问题排查）",
                    "type": "string",
                    "example": "trace-1729756800-a1b2c3d4"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "输入 \"Bearer {token}\" 格式的 JWT 令牌进行身份认证",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}